requirements.txt
torch
torchvision
matplotlib
pyyaml
opencv-python
tqdm
h5py
tensorboard
wandb
pandas
